+++ IcVibracoes.Core/Calculator/MainMatrixes/Beam/Circular/CircularBeamMainMatrix.cs
@@ -5,7 +5,5 @@ namespace IcVibracoes.Core.Calculator.MainMatrixes.Beam.Circular
     /// <summary>
     /// It's responsible to calculate the circular beam main matrixes.
     /// </summary>
-    public class CircularBeamMainMatrix : BeamMainMatrix<CircularProfile>, ICircularBeamMainMatrix
-    {
-    }
+    public class CircularBeamMainMatrix : BeamMainMatrix<CircularProfile>, ICircularBeamMainMatrix { }
 }