+++ IcVibracoes.Core/Calculator/MainMatrixes/Beam/Circular/ICircularBeamMainMatrix.cs
@@ -2,7 +2,8 @@
 
 namespace IcVibracoes.Core.Calculator.MainMatrixes.Beam.Circular
 {
-    public interface ICircularBeamMainMatrix : IBeamMainMatrix<CircularProfile>
-    {
-    }
+    /// <summary>
+    /// It's responsible to calculate the circular beam main matrixes.
+    /// </summary>
+    public interface ICircularBeamMainMatrix : IBeamMainMatrix<CircularProfile> { }
 }