+++ IcVibracoes.Core/Calculator/MainMatrixes/BeamWithPiezoelectric/Rectangular/IRectangularBeamWithPiezoelectricMainMatrix.cs
@@ -5,7 +5,5 @@ namespace IcVibracoes.Core.Calculator.MainMatrixes.BeamWithPiezoelectric.Rectang
     /// <summary>
     /// It's responsible to calculate the rectangular beam with piezoelectric main matrixes.
     /// </summary>
-    public interface IRectangularBeamWithPiezoelectricMainMatrix : IBeamWithPiezoelectricMainMatrix<RectangularProfile>
-    {
-    }
+    public interface IRectangularBeamWithPiezoelectricMainMatrix : IBeamWithPiezoelectricMainMatrix<RectangularProfile> { }
 }